Refactor Filament side
