+++ app-modules/engagement/resources/views/filament/actions/draft-template-with-ai-modal-content.blade.php
