+++ app-modules/engagement/src/Filament/Actions/RelationManagerDraftWithAiAction.php
@@ -34,7 +34,7 @@
 </COPYRIGHT>
 */
 
-namespace AdvisingApp\Engagement\Filament\Actions;
+namespace AdvisingApp\Engagement\Filament\ManageRelatedRecords\ManageRelatedEngagementRecords\Actions;
 
 use AdvisingApp\Ai\Actions\CompletePrompt;
 use AdvisingApp\Ai\Exceptions\MessageResponseException;
@@ -49,12 +49,13 @@ use Filament\Forms\Components\Textarea;
 use Filament\Forms\Get;
 use Filament\Forms\Set;
 use Filament\Notifications\Notification;
-use Filament\Resources\Pages\Page;
+use Filament\Resources\Pages\ManageRelatedRecords;
+use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Support\Enums\MaxWidth;
 use Illuminate\Support\Facades\Vite;
 use Illuminate\Support\Str;
 
-class DraftWithAiAction extends Action
+class RelationManagerDraftWithAiAction extends Action
 {
     protected array | Closure $mergeTags = [];
 
@@ -66,8 +67,8 @@ class DraftWithAiAction extends Action
             ->label('Draft with AI Assistant')
             ->link()
             ->icon('heroicon-m-pencil')
-            ->modalContent(fn (Page $livewire) => view('engagement::filament.actions.draft-with-ai-modal-content', [
-                'recordTitle' => $livewire::getResource()::getPluralModelLabel(),
+            ->modalContent(fn (RelationManager $livewire) => view('engagement::filament.actions.draft-with-ai-modal-content', [
+                'recordTitle' => $livewire->getOwnerRecord()->getAttribute($livewire->getOwnerRecord()::displayNameKey()),
                 'avatarUrl' => AiAssistant::query()->where('is_default', true)->first()
                     ?->getFirstTemporaryUrl(now()->addHour(), 'avatar', 'avatar-height-250px') ?: Vite::asset('resources/images/canyon-ai-headshot.jpg'),
             ]))
@@ -80,13 +81,13 @@ class DraftWithAiAction extends Action
                     ->placeholder('What do you want to write about?')
                     ->required(),
             ])
-            ->action(function (array $data, Get $get, Set $set, Page $livewire) {
+            ->action(function (array $data, Get $get, Set $set, ManageRelatedRecords | RelationManager $livewire) {
                 $model = app(AiIntegratedAssistantSettings::class)->default_model;
 
                 $userName = auth()->user()->name;
                 $userJobTitle = auth()->user()->job_title ?? 'staff member';
                 $clientName = app(LicenseSettings::class)->data->subscription->clientName;
-                $educatableLabel = $livewire::getResource()::getPluralModelLabel();
+                $educatableLabel = $livewire->getOwnerRecord()::getLabel();
 
                 $mergeTagsList = collect($this->getMergeTags())
                     ->map(fn (string $tag): string => <<<HTML
@@ -100,7 +101,7 @@ class DraftWithAiAction extends Action
                             aiModel: $model,
                             prompt: <<<EOL
                                 The user's name is {$userName} and they are a {$userJobTitle} at {$clientName}.
-                                Please draft a short SMS message for {$educatableLabel} at their college.
+                                Please draft a short SMS message for a {$educatableLabel} at their college.
                                 The user will send a message to you containing instructions for the content.
 
                                 You should only respond with the SMS content, you should never greet them.
@@ -134,7 +135,7 @@ class DraftWithAiAction extends Action
                         aiModel: $model,
                         prompt: <<<EOL
                             The user's name is {$userName} and they are a {$userJobTitle} at {$clientName}.
-                            Please draft an email for {$educatableLabel} at their college.
+                            Please draft an email for a {$educatableLabel} at their college.
                             The user will send a message to you containing instructions for the content.
 
                             You should only respond with the email content, you should never greet them.
