+++ app-modules/engagement/src/Filament/Resources/EngagementResponseResource/Actions/EngagementResponseViewAction.php
@@ -1,4 +1,6 @@
-{{--
+<?php
+
+/*
 <COPYRIGHT>
 
     Copyright © 2016-2025, Canyon GBS LLC. All rights reserved.
@@ -30,31 +32,22 @@
     https://www.canyongbs.com or contact us via email at legal@canyongbs.com.
 
 </COPYRIGHT>
---}}
-<div class="mb-3 flex gap-4 text-base md:gap-6">
-    <div class="flex flex-shrink-0 flex-col items-end">
-        <img
-            class="h-8 w-8 rounded-full object-cover object-center"
-            src="{{ $avatarUrl }}"
-            alt="Assistant avatar"
-        >
-    </div>
+*/
 
-    <div class="prose h-36 flex-1 dark:prose-invert sm:h-20">
-        <p
-            x-data="{ content: '' }"
-            x-init="const message = @js('Hi ' . auth()->user()->name . ", I am happy to help you draft this message to {$recordTitle}. Please describe what you would like the proposed message to say:");
-            
-            const typeWord = async (word, delay) => {
-                content += word + ' ';
-            
-                await new Promise(resolve => setTimeout(resolve, delay));
-            };
-            
-            for (const word of message.split(' ')) {
-                await typeWord(word, Math.floor(Math.random() * 100));
-            }"
-            x-text="content"
-        ></p>
-    </div>
-</div>
+namespace AdvisingApp\Engagement\Filament\Resources\EngagementResponseResource\Actions;
+
+use Filament\Actions\ViewAction;
+use Filament\Infolists\Components\TextEntry;
+
+class EngagementResponseViewAction
+{
+    public static function make(): ViewAction
+    {
+        return ViewAction::make()
+            ->infolist([
+                TextEntry::make('content'),
+                TextEntry::make('sent_at')
+                    ->dateTime('Y-m-d H:i:s'),
+            ]);
+    }
+}
